# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Router",
    "version": "generated"
  },
  "paths": {
...
    "/api/v1/route": {
      "post": {
        "tags": [
          "routing"
        ],
        "summary": "Route",
        "operationId": "route_api_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/route/simulate": {
      "post": {
        "tags": [
          "routing"
        ],
        "summary": "Simulate",
        "description": "Route AS `agentId` (its live scope and budgets) and explain the result.\nNever publishes exposure to any session, never executes, never touches\nthe route cache; the decision row is marked `simulated/`.",
        "operationId": "simulate_api_v1_route_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/route/evaluate": {
      "post": {
        "tags": [
          "routing"
        ],
        "summary": "Evaluate",
        "description": "Run a named, server-shipped dataset against the LIVE pipeline + scope\nresolver and store the result. Dataset names are whitelisted (no paths).",
        "operationId": "evaluate_api_v1_route_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluateBody": {
        "properties": {
          "dataset": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Dataset"
          },
          "maxTools": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Maxtools",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "dataset"
        ],
        "title": "EvaluateBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RouteBody": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 4000,
            "title": "Query"
          },
          "agentId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentid"
          },
          "maxTools": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxtools"
          },
          "maxServers": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxservers"
          },
          "allowedServers": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 120,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowedservers"
          },
          "maxSkills": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxskills"
          },
          "kinds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "tool",
                    "skill"
                  ]
                },
                "type": "array",
                "maxItems": 2,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Kinds"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RouteBody"
      },
      "RouteResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/RoutedToolOut"
            },
            "type": "array",
            "title": "Tools"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/RoutedSkillOut"
            },
            "type": "array",
            "title": "Skills"
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          },
          "no_match": {
            "type": "boolean",
            "title": "No Match"
          },
          "max_tools_applied": {
            "type": "integer",
            "title": "Max Tools Applied"
          },
          "max_servers_applied": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Servers Applied"
          },
          "max_skills_applied": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Skills Applied"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "tools",
          "skills",
          "fallback_used",
          "latency_ms",
          "no_match",
          "max_tools_applied",
          "max_servers_applied",
          "max_skills_applied",
          "cached"
        ],
        "title": "RouteResponse"
      },
...
      "RoutedToolOut": {
        "properties": {
          "server": {
            "type": "string",
            "title": "Server"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "server",
          "tool",
          "score"
        ],
        "title": "RoutedToolOut"
      },
...
      "SimulateBody": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 4000,
            "title": "Query"
          },
          "agentId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Agentid"
          },
          "maxTools": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxtools"
          },
          "maxServers": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxservers"
          },
          "allowedServers": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 120,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowedservers"
          },
          "maxSkills": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxskills"
          },
          "kinds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "tool",
                    "skill"
                  ]
                },
                "type": "array",
                "maxItems": 2,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Kinds"
          }
        },
        "type": "object",
        "required": [
          "query",
          "agentId"
        ],
        "title": "SimulateBody"
      },
      "SimulateResponse": {
        "properties": {
          "requestId": {
            "type": "string",
            "title": "Requestid"
          },
          "agentId": {
            "type": "string",
            "title": "Agentid"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/SimulatedToolOut"
            },
            "type": "array",
            "title": "Tools"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SimulatedSkillOut"
            },
            "type": "array",
            "title": "Skills"
          },
          "noMatch": {
            "type": "boolean",
            "title": "Nomatch"
          },
          "fallbackUsed": {
            "type": "boolean",
            "title": "Fallbackused"
          },
          "latencyMs": {
            "type": "number",
            "title": "Latencyms"
          },
          "modelVersion": {
            "type": "string",
            "title": "Modelversion"
          },
          "maxToolsApplied": {
            "type": "integer",
            "title": "Maxtoolsapplied"
          },
          "maxServersApplied": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxserversapplied"
          },
          "maxSkillsApplied": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxskillsapplied"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/DiagnosticsOut"
          }
        },
        "type": "object",
        "required": [
          "requestId",
          "agentId",
          "simulated",
          "tools",
          "skills",
          "noMatch",
          "fallbackUsed",
          "latencyMs",
          "modelVersion",
          "maxToolsApplied",
          "maxServersApplied",
          "maxSkillsApplied",
          "diagnostics"
        ],
        "title": "SimulateResponse"
      },
...
    }
  }
}